Improve font display in MacVim
It turns out Vim was using Hack as my font before. I was intending to
use Fira Code, which may be what it is now. Not sure if it'll work for
startup, but a simple `:so ~/.vim/vimrc` seems to do it. I've also
turned on macligatures, so that we get ridiculous obnoxious symbols when
we do things like www or <$>, or 0xFF.

       # modified:   vim/vimrc